+++ python/core/asset.py
@@ -492,3 +492,33 @@ class Asset(WorkdirEnabled):
             self.asset_dict['use_path_as_workpath'] = 1
         else:
             self.asset_dict['use_path_as_workpath'] = 0
+
+
+class NorefAsset(Asset):
+    """
+    NorefAsset is similar to Asset except that it does not have a reference
+    video path ref_path.
+    """
+
+    # ==== constructor ====
+
+    def __init__(self, dataset, content_id, asset_id,
+                 dis_path,
+                 asset_dict,
+                 workdir_root=config.ROOT + "/workspace/workdir"):
+        """
+        :param dataset
+        :param content_id: ID of content the asset correspond to within dataset
+        :param asset_id: ID of asset
+        :param dis_path: path to distorted video
+        :param asset_dict: dictionary with additional asset properties
+        :param workdir_root:
+        :return:
+        """
+        super(Asset, self).__init__(workdir_root)
+        self.dataset = dataset
+        self.content_id = content_id
+        self.asset_id = asset_id
+        self.ref_path = dis_path # just assign ref_path same as dis_path
+        self.dis_path = dis_path
+        self.asset_dict = asset_dict