+++ CHANGELOG.md
@@ -12,4 +12,10 @@
 ## [1.0.2]
 
 **New features:**
-- Refactor feature code to expose ssim/ms-ssim, speed up ssim/ms-ssim
+- Refactor feature code to expose ssim/ms-ssim, speed up ssim/ms-ssim.
+
+## [1.0.3]
+
+**New features:**
+- Add base class NorefFeatureExtractor for any feature extractor that do not use a reference video.
+- Add MomentNorefFeatureExtractor subclassing NorefFeatureExtractor as an example implementation.